backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
openai==1.35.0
elevenlabs==1.2.0
moviepy==1.0.3
Pillow==10.3.0
python-dotenv==1.0.1
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
aiofiles==23.2.1
requests==2.32.3
